+++ app/src/main/java/com/credman/cmwallet/getcred/GetCredentialActivity.kt
@@ -9,10 +9,8 @@ import androidx.biometric.BiometricManager
 import androidx.biometric.BiometricPrompt
 import androidx.credentials.CreateCredentialRequest
 import androidx.credentials.CreateCustomCredentialRequest
-import androidx.credentials.CustomCredential
 import androidx.credentials.DigitalCredential
 import androidx.credentials.ExperimentalDigitalCredentialApi
-import androidx.credentials.GetCredentialResponse
 import androidx.credentials.GetDigitalCredentialOption
 import androidx.credentials.exceptions.GetCredentialUnknownException
 import androidx.credentials.provider.PendingIntentHandler
@@ -28,7 +26,6 @@ import com.credman.cmwallet.data.model.CredentialItem
 import com.credman.cmwallet.data.model.CredentialKeySoftware
 import com.credman.cmwallet.decodeBase64UrlNoPadding
 import com.credman.cmwallet.getcred.GetCredentialActivity.DigitalCredentialResult
-import com.credman.cmwallet.intToBigEndianByteArray
 import com.credman.cmwallet.loadECPrivateKey
 import com.credman.cmwallet.mdoc.createSessionTranscript
 import com.credman.cmwallet.mdoc.filterIssuerSigned
@@ -40,27 +37,10 @@ import com.credman.cmwallet.openid4vp.OpenId4VP
 import com.credman.cmwallet.openid4vp.OpenId4VPMatchedCredential
 import com.credman.cmwallet.openid4vp.OpenId4VPMatchedMDocClaims
 import com.credman.cmwallet.toBase64UrlNoPadding
-import com.credman.cmwallet.toJWK
 import com.google.android.gms.identitycredentials.IntentHelper
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.Json
-import org.jose4j.jwe.kdf.ConcatKeyDerivationFunction
 import org.json.JSONObject
-import java.math.BigInteger
-import java.nio.ByteBuffer
-import java.nio.ByteOrder
-import java.security.AlgorithmParameters
-import java.security.KeyFactory
-import java.security.KeyPairGenerator
-import java.security.SecureRandom
-import java.security.spec.ECGenParameterSpec
-import java.security.spec.ECParameterSpec
-import java.security.spec.ECPoint
-import java.security.spec.ECPublicKeySpec
-import javax.crypto.Cipher
-import javax.crypto.KeyAgreement
-import javax.crypto.spec.GCMParameterSpec
-import javax.crypto.spec.SecretKeySpec
 
 
 fun createOpenID4VPResponse(
@@ -127,92 +107,7 @@ fun createOpenID4VPResponse(
     }
 
     // Create the openid4vp result
-    val responseJson = JSONObject().put("vp_token", vpToken).toString()
-    val response = if (openId4VPRequest.responseMode == "dc_api.jwt") {
-        // Encrypt response if applicable
-        val encryptionAgl = openId4VPRequest.clientMedtadata?.opt("authorization_encrypted_response_alg")
-        val encryptionEnc = openId4VPRequest.clientMedtadata?.opt("authorization_encrypted_response_enc")
-        val signAgl = openId4VPRequest.clientMedtadata?.opt("authorization_signed_response_alg")
-        val jwks = openId4VPRequest.clientMedtadata?.opt("jwks")
-        if (encryptionAgl != null && encryptionEnc != null && signAgl == null) {
-            require(encryptionAgl == "ECDH-ES" && encryptionEnc == "A128GCM") { "Unsupported encryption algorithm" }
-            val jwks = (jwks!! as JSONObject).getJSONArray("keys")
-            var encryptionJwk = jwks[0] as JSONObject
-            for (i in 0..<jwks.length()) {
-                val jwk = jwks[i] as JSONObject
-                if (jwk.has("use")
-                    && jwk["use"] == "enc"
-                    && encryptionJwk["kty"] == "EC"
-                    && encryptionJwk["crv"] == "P-256"
-                ) {
-                    encryptionJwk = jwk
-                }
-            }
-            val kid = encryptionJwk.optString("kid")
-            val x = encryptionJwk.getString("x")
-            val y = encryptionJwk.getString("y")
-            val kf = KeyFactory.getInstance("EC")
-            val parameters = AlgorithmParameters.getInstance("EC")
-            parameters.init(ECGenParameterSpec("secp256r1"))
-            val publicKey = kf.generatePublic(
-                ECPublicKeySpec(
-                    ECPoint(
-                        BigInteger(1, x.decodeBase64UrlNoPadding()),
-                        BigInteger(1, y.decodeBase64UrlNoPadding())
-                    ),
-                    parameters.getParameterSpec(ECParameterSpec::class.java)
-                )
-            )
-            val kpg =  KeyPairGenerator.getInstance("EC")
-            kpg.initialize(ECGenParameterSpec("secp256r1"))
-            val kp = kpg.genKeyPair()
-            val partyUInfo = ByteArray(0)
-            val partyVInfo = ByteArray(0)
-            val header = JSONObject()
-            header.put("apu", partyUInfo.toBase64UrlNoPadding())
-            header.put("apv", partyVInfo.toBase64UrlNoPadding())
-            header.put("alg", "ECDH-ES")
-            header.put("enc", "A128GCM")
-            header.put("epk", JSONObject(kp.public.toJWK().toString()))
-            val headerEncoded = header.toString().toByteArray().toBase64UrlNoPadding()
-
-            val keyAgreement = KeyAgreement.getInstance("ECDH")
-            keyAgreement.init(kp.private)
-            keyAgreement.doPhase(publicKey, true)
-            val sharedSecret = keyAgreement.generateSecret()
-            val concatKdf = ConcatKeyDerivationFunction("SHA-256")
-
-            val algOctets = (encryptionEnc as String).toByteArray()
-            val keydatalen = 128
-
-            val derivedKey = concatKdf.kdf(
-                sharedSecret,
-                keydatalen,
-                intToBigEndianByteArray(algOctets.size) + algOctets,
-                intToBigEndianByteArray(partyUInfo.size) + partyUInfo,
-                intToBigEndianByteArray(partyVInfo.size) + partyVInfo,
-                intToBigEndianByteArray(keydatalen),
-                ByteArray(0)
-            )
-            val sks = SecretKeySpec(derivedKey, "AES")
-            val aesCipher = Cipher.getInstance("AES/GCM/NoPadding")
-            val iv = ByteArray(12)
-            SecureRandom().nextBytes(iv)
-            val ivEncoded = iv.toBase64UrlNoPadding()
-            aesCipher.init(Cipher.ENCRYPT_MODE, sks, GCMParameterSpec(128, iv))
-            aesCipher.updateAAD(headerEncoded.toByteArray())
-            val encrypted = aesCipher.doFinal(responseJson.toString().toByteArray())
-            val ct = encrypted.slice(0 until (encrypted.size - 16)).toByteArray()
-            val ctEncoded = ct.toBase64UrlNoPadding()
-            val tag = encrypted.slice((encrypted.size - 16) until encrypted.size).toByteArray()
-            val tagEncoded = tag.toBase64UrlNoPadding()
-            "${headerEncoded}..${ivEncoded}.${ctEncoded}.${tagEncoded}"
-        } else {
-            throw UnsupportedOperationException("Response should be signed and / or encrypted but it's not supported yet")
-        }
-    } else {
-        responseJson
-    }
+    val response = openId4VPRequest.generateResponse(vpToken)
     Log.d(TAG, "Returning $response")
     return DigitalCredentialResult(
         responseJson = response,
