+++ app/src/main/java/com/credman/cmwallet/openid4vp/OpenId4VP.kt
@@ -1,9 +1,9 @@
 package com.credman.cmwallet.openid4vp
 
 import android.util.Base64
-import com.credman.cmwallet.cbor.CborTag
 import com.credman.cmwallet.cbor.cborEncode
 import com.credman.cmwallet.data.model.CredentialItem
+import com.credman.cmwallet.jweSerialization
 import org.json.JSONObject
 import java.security.MessageDigest
 
@@ -143,6 +143,38 @@ class OpenId4VP(val request: String, val clientId: String) {
             oid4vpHandoverIdentifier,
             md.digest(cborEncode(handoverData))
         )
+    }
+
+    fun generateResponse(vpToken: JSONObject): String {
+        val responseJson = JSONObject().put("vp_token", vpToken).toString()
+        val response = if (responseMode == "dc_api.jwt") {
+            // Encrypt response if applicable
+            val encryptionAgl = clientMedtadata?.opt("authorization_encrypted_response_alg")
+            val encryptionEnc = clientMedtadata?.opt("authorization_encrypted_response_enc")
+            val signAgl = clientMedtadata?.opt("authorization_signed_response_alg")
+            val jwks = clientMedtadata?.opt("jwks")
+            if (encryptionAgl != null && encryptionEnc != null && signAgl == null) {
+                require(encryptionAgl == "ECDH-ES" && encryptionEnc == "A128GCM") { "Unsupported encryption algorithm" }
+                val jwks = (jwks!! as JSONObject).getJSONArray("keys")
+                var encryptionJwk = jwks[0] as JSONObject
+                for (i in 0..<jwks.length()) {
+                    val jwk = jwks[i] as JSONObject
+                    if (jwk.has("use")
+                        && jwk["use"] == "enc"
+                        && encryptionJwk["kty"] == "EC"
+                        && encryptionJwk["crv"] == "P-256"
+                    ) {
+                        encryptionJwk = jwk
+                    }
+                }
+                jweSerialization(encryptionJwk, responseJson)
+            } else {
+                throw UnsupportedOperationException("Response should be signed and / or encrypted but it's not supported yet")
+            }
+        } else {
+            responseJson
+        }
+        return response
     }
 
     companion object {
